Dispatch Errors on Main Thread
The other error handlers are called by Starscream's delegate, but the auth
errors are called by NSUrlSession, which is on a background thread.

So, I've moved the logic into a dispatch_async the main thread.
I decided to put it in the handleError method rather than outside it because
there are multiple places that need it and currently all of them are on the
background threads.

If later that's not true, then we should put the dispatch around the method
call in all 4 or so places it comes up.