Reconnect on Disconnect Unless Reachability Changes
If the app is allowed to go to sleep in the background, or perhaps even if the
websocket is just closed from the remote end, it will never reconnect.

We're currently relying on the reachability to tell us when it's time to
connect again, but that's assuming that the reason we disconnected was
reachability.

That's not always the case.
So, I've started a timer on disconnect that gives reachability 1 second to
figure itself out before we try to reconnect.
We cancel the operation if reachability changes, otherwise we let it run.