Don't Reconnect If Network is Not Reachable
This is to cover the case where the app starts in a state of unreachability.
We will have already been told that the network was unreachable before we get
to the websocket failure, so don't bother retrying in the case where we know
the network is unreachable.

In the worst case, even if the network is reachable, we know that if the system
doesn't know that it is, we will receive the notification when the system
realizes it, and that will cause us to reconnect.